Factor common functionality in QueryBuilder
