+++ src/backend/mod.rs
@@ -19,72 +19,13 @@ pub use postgres::*;
 #[cfg(feature="backend-sqlite")]
 pub use sqlite::*;
 
+mod query_builder;
+
+pub use self::query_builder::*;
+
 pub trait GenericBuilder: QueryBuilder + SchemaBuilder {}
 
 pub trait SchemaBuilder: TableBuilder + IndexBuilder + ForeignKeyBuilder {}
-
-pub trait QueryBuilder {
-
-    /// Translate [`InsertStatement`] into SQL statement.
-    fn prepare_insert_statement(&self, insert: &InsertStatement, sql: &mut SqlWriter, collector: &mut dyn FnMut(Value));
-
-    /// Translate [`SelectStatement`] into SQL statement.
-    fn prepare_select_statement(&self, select: &SelectStatement, sql: &mut SqlWriter, collector: &mut dyn FnMut(Value));
-
-    /// Translate [`UpdateStatement`] into SQL statement.
-    fn prepare_update_statement(&self, update: &UpdateStatement, sql: &mut SqlWriter, collector: &mut dyn FnMut(Value));
-
-    /// Translate [`DeleteStatement`] into SQL statement.
-    fn prepare_delete_statement(&self, delete: &DeleteStatement, sql: &mut SqlWriter, collector: &mut dyn FnMut(Value));
-
-    /// Translate [`SimpleExpr`] into SQL statement.
-    fn prepare_simple_expr(&self, simple_expr: &SimpleExpr, sql: &mut SqlWriter, collector: &mut dyn FnMut(Value));
-
-    /// Translate [`SelectDistinct`] into SQL statement.
-    fn prepare_select_distinct(&self, select_distinct: &SelectDistinct, sql: &mut SqlWriter, collector: &mut dyn FnMut(Value));
-
-    /// Translate [`SelectExpr`] into SQL statement.
-    fn prepare_select_expr(&self, select_expr: &SelectExpr, sql: &mut SqlWriter, collector: &mut dyn FnMut(Value));
-
-    /// Translate [`JoinExpr`] into SQL statement.
-    fn prepare_join_expr(&self, join_expr: &JoinExpr, sql: &mut SqlWriter, collector: &mut dyn FnMut(Value));
-
-    /// Translate [`TableRef`] into SQL statement.
-    fn prepare_table_ref(&self, table_ref: &TableRef, sql: &mut SqlWriter, collector: &mut dyn FnMut(Value));
-
-    /// Translate [`UnOper`] into SQL statement.
-    fn prepare_un_oper(&self, un_oper: &UnOper, sql: &mut SqlWriter, collector: &mut dyn FnMut(Value));
-
-    /// Translate [`BinOper`] into SQL statement.
-    fn prepare_bin_oper(&self, bin_oper: &BinOper, sql: &mut SqlWriter, collector: &mut dyn FnMut(Value));
-
-    /// Translate [`LogicalChainOper`] into SQL statement.
-    fn prepare_logical_chain_oper(&self, log_chain_oper: &LogicalChainOper, i: usize, length: usize, sql: &mut SqlWriter, collector: &mut dyn FnMut(Value));
-
-    /// Translate [`Function`] into SQL statement.
-    fn prepare_function(&self, function: &Function, sql: &mut SqlWriter, collector: &mut dyn FnMut(Value));
-
-    /// Translate [`JoinType`] into SQL statement.
-    fn prepare_join_type(&self, join_type: &JoinType, sql: &mut SqlWriter, collector: &mut dyn FnMut(Value));
-
-    /// Translate [`OrderExpr`] into SQL statement.
-    fn prepare_order_expr(&self, order_expr: &OrderExpr, sql: &mut SqlWriter, collector: &mut dyn FnMut(Value));
-
-    /// Translate [`JoinOn`] into SQL statement.
-    fn prepare_join_on(&self, join_on: &JoinOn, sql: &mut SqlWriter, collector: &mut dyn FnMut(Value));
-
-    /// Translate [`Order`] into SQL statement.
-    fn prepare_order(&self, order: &Order, sql: &mut SqlWriter, collector: &mut dyn FnMut(Value));
-
-    /// Translate [`Value`] into SQL statement.
-    fn prepare_value(&self, value: &Value, sql: &mut SqlWriter, collector: &mut dyn FnMut(Value));
-
-    /// Translate [`Keyword`] into SQL statement.
-    fn prepare_keyword(&self, keyword: &Keyword, sql: &mut SqlWriter, collector: &mut dyn FnMut(Value));
-
-    /// Convert a SQL value into syntax-specific string
-    fn value_to_string(&self, v: &Value) -> String;
-}
 
 pub trait TableBuilder {
