+++ src/backend/postgres/table.rs
@@ -1,5 +1,4 @@
 use super::*;
-use super::query::pg_value_to_string;
 
 impl TableBuilder for PostgresQueryBuilder {
     fn prepare_table_create_statement(&self, create: &TableCreateStatement, sql: &mut SqlWriter) {
@@ -133,7 +132,7 @@ impl TableBuilder for PostgresQueryBuilder {
         match column_spec {
             ColumnSpec::Null => write!(sql, "NULL"),
             ColumnSpec::NotNull => write!(sql, "NOT NULL"),
-            ColumnSpec::Default(value) => write!(sql, "DEFAULT {}", pg_value_to_string(value)),
+            ColumnSpec::Default(value) => write!(sql, "DEFAULT {}", self.value_to_string(value)),
             ColumnSpec::AutoIncrement => write!(sql, ""),
             ColumnSpec::UniqueKey => write!(sql, "UNIQUE"),
             ColumnSpec::PrimaryKey => write!(sql, "PRIMARY KEY"),
