+++ src/backend/sqlite/table.rs
@@ -1,5 +1,4 @@
 use super::*;
-use super::query::sqlite_value_to_string;
 
 impl TableBuilder for SqliteQueryBuilder {
     fn prepare_table_create_statement(&self, create: &TableCreateStatement, sql: &mut SqlWriter) {
@@ -153,7 +152,7 @@ impl TableBuilder for SqliteQueryBuilder {
         match column_spec {
             ColumnSpec::Null => write!(sql, "NULL"),
             ColumnSpec::NotNull => write!(sql, "NOT NULL"),
-            ColumnSpec::Default(value) => write!(sql, "DEFAULT {}", sqlite_value_to_string(value)),
+            ColumnSpec::Default(value) => write!(sql, "DEFAULT {}", self.value_to_string(value)),
             ColumnSpec::AutoIncrement => write!(sql, "AUTOINCREMENT"),
             ColumnSpec::UniqueKey => write!(sql, "UNIQUE"),
             ColumnSpec::PrimaryKey => write!(sql, "PRIMARY KEY"),
